Fix #800
